Refactor Feed and Post to use redux.
